Upgraded to Play Framework 2.5.0
Version 1.2.0.

This was painful!  Play now deprecates global state in favour of dependency injection at runtime.  Tradtunestore relied heavily on a singleton object Utils which was available to all components and contained configuration state and other utility functions.  It is not possible to inject configuration  state into a global object and so a fair amount of refactoring was necessary to inject the configuration state into the Controller and then to pass that state to other classes which required the config.  This in turn means that Proxy, too, becomes a class and not an object and the login form (which also requires config) is now embedded in a new class.  This, of course, considerably complicates the code just in order to retain the same functionality.

Also integrated the interactive ABC tutorial (written in elm but integrated as javascript).