Integrated the elm ABC Tutorial
Amended the interface to main.scala.html to allow for a scripts parameter defaulted to empty HTML.  The ABC Editor and ABC Tutorial populate this scripts variable with the various pieces of elm-generated and/or hand-crafted javascript needed to implement these pages.

It may be possible to embed the javascript for the main player for uploaded tunes in this manner but I haven't yet attempted to do so.

Added to the help file accordingly.